Introduce a Transform step to the parquet writer Flush, to transform the chunks of a rowgroup into the shape that the column writers require
