+++ extension/parquet/writer/CMakeLists.txt
@@ -9,6 +9,9 @@ add_library_unity(
   primitive_column_writer.cpp
   variant_column_writer.cpp
   struct_column_writer.cpp)
+
+add_subdirectory(variant)
+
 set(PARQUET_EXTENSION_FILES
     ${PARQUET_EXTENSION_FILES} $<TARGET_OBJECTS:duckdb_parquet_writers>
     PARENT_SCOPE)
